pyproject.toml
[project]
name = "test-mkk-luna"
version = "0.1.0"
description = "Test task for MKK Luna"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "geoalchemy2>=0.18.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
]


[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio", "httpx"]

[tool.ruff]
line-length = 88
target-version = "py310"  
per-file-ignores = { "__init__.py" = ["F401"] }
exclude = [
    "__init__",
    ".bzr",
    ".direnv", 
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "A"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
